backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.1.7
django-cors-headers==4.7.0
django-ninja==1.3.0
djangorestframework_simplejwt==5.5.0
dramatiq==1.17.1
prometheus_client==0.21.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
django_dramatiq==0.13.0
redis==5.2.1